pyproject.toml
[project]
name = "graphius"
version = "0.1.0"
description = "A Simple Graph Library with Visualization Support"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Alperen ÜNLÜ"}
]
requires-python = ">=3.10"
dependencies = [
    "graphviz>=0.21",
    "pillow>=11.3.0",
]

[project.urls]
Source = "https://github.com/alperenunlu/graphius"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
publish-url = "https://pypi.org/legacy/"
